refactor: renaming classes
