+++ lib/RuntimeClientFactory/index.ts
@@ -1,27 +1,27 @@
 import { State } from '@voiceflow/runtime';
 
-import Agent from '@/lib/Agent';
+import RuntimeClient from '@/lib/RuntimeClient';
 import Client from '@/lib/Client';
 import { DataConfig } from '@/lib/types';
 
 import { validateVarMerge } from '../Variables/utils';
 import { DEFAULT_ENDPOINT } from './constants';
 
-export type AppConfig<S extends State['variables']> = {
+export type FactoryConfig<S extends State['variables']> = {
   versionID: string;
   endpoint?: string;
   dataConfig?: DataConfig;
   variables?: Partial<S>;
 };
 
-export class App<S extends Record<string, any> = Record<string, any>> {
+export class RuntimeClientFactory<S extends Record<string, any> = Record<string, any>> {
   public client: Client;
 
   private dataConfig: DataConfig;
 
   private defaultState: State;
 
-  constructor({ versionID, endpoint = DEFAULT_ENDPOINT, dataConfig, variables }: AppConfig<S>) {
+  constructor({ versionID, endpoint = DEFAULT_ENDPOINT, dataConfig, variables }: FactoryConfig<S>) {
     if (variables) {
       validateVarMerge(variables);
     }
@@ -37,9 +37,9 @@ export class App<S extends Record<string, any> = Record<string, any>> {
     };
   }
 
-  createAgent(state: State = this.defaultState) {
-    return new Agent(state, { client: this.client, dataConfig: this.dataConfig });
+  createClient(state: State = this.defaultState) {
+    return new RuntimeClient(state, { client: this.client, dataConfig: this.dataConfig });
   }
 }
 
-export default App;
+export default RuntimeClientFactory;
